requirements.txt
langchain>=0.0.27
langchain-community>=0.0.3
python-dotenv>=0.19.0
openai>=1.3.0 